Apply code reviews
