+++ site/src/pages/TaskPage/TaskAppIframe.tsx
@@ -9,8 +9,8 @@ import {
 import { Spinner } from "components/Spinner/Spinner";
 import { useProxy } from "contexts/ProxyContext";
 import { EllipsisVertical, ExternalLinkIcon, HouseIcon } from "lucide-react";
-import type { WorkspaceAppWithAgent } from "modules/apps/apps";
 import { useAppLink } from "modules/apps/useAppLink";
+import type { WorkspaceAppWithAgent } from "modules/tasks/apps";
 import { type FC, type HTMLProps, useRef } from "react";
 import { Link as RouterLink } from "react-router";
 import { cn } from "utils/cn";
