+++ site/src/pages/TaskPage/TaskApps.tsx
@@ -11,12 +11,12 @@ import { InfoTooltip } from "components/InfoTooltip/InfoTooltip";
 import { Link } from "components/Link/Link";
 import { ScrollArea, ScrollBar } from "components/ScrollArea/ScrollArea";
 import { ChevronDownIcon, LayoutGridIcon, TerminalIcon } from "lucide-react";
+import { getTerminalHref } from "modules/apps/apps";
+import { useAppLink } from "modules/apps/useAppLink";
 import {
 	getAllAppsWithAgent,
-	getTerminalHref,
 	type WorkspaceAppWithAgent,
-} from "modules/apps/apps";
-import { useAppLink } from "modules/apps/useAppLink";
+} from "modules/tasks/apps";
 import { type FC, useState } from "react";
 import { type LinkProps, Link as RouterLink } from "react-router";
 import { cn } from "utils/cn";
